Limit PartialEq to tests for secrets
