+++ src/tree/private_tree.rs
@@ -9,7 +9,8 @@ use crate::utils::zero;
 
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(test, derive(PartialEq))]
 pub(crate) struct CommitSecret {
     secret: Secret,
 }
@@ -38,7 +39,8 @@ impl CommitSecret {
 
 pub(crate) type PathSecrets = Vec<Secret>;
 
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(test, derive(PartialEq))]
 pub(crate) struct PrivateTree {
     // The index of the node corresponding to this leaf information.
     node_index: NodeIndex,
